requirements.txt
torch>=2.4.0 --index-url https://download.pytorch.org/whl/cu118
numpy
datasets
matplotlib
pandas
requests
scikit-learn
tqdm
tokenizers
tiktoken
pytorch-lightning
